Extract god-mode parsing into PlanetfallGodModeCommands helper; apply review polish
PlanetfallContext was accumulating ~90 lines of god-mode command parsing,
crowding out its real job of holding game state. The parsing now lives in
Planetfall/Command/PlanetfallGodModeCommands (alongside BedCommands and
DeathProcessor); the context keeps only the one-line IGodModeCommandHandler
delegation and the serialized flags.

Also applies the code-review suggestions: a comment marking the sub-handler
evaluation order as deliberate (reset must win "reset sleep clock" over the
survival toggle), a response-text assertion on the wandering re-enable test,
and an IGodModeCommandHandler doc note that input includes the "god mode"
prefix so implementations match on trigger words, not position.
